requirements.txt
lxml
python-dotenv
tkinter